requirements.txt
streamlit==1.41.1
pandas==2.2.3
openpyxl==3.1.5
httpx==0.28.1
beautifulsoup4==4.12.3
trafilatura==1.12.2
openai>=1.40.0
python-dotenv
